requirements.txt
absl-py==2.1.0
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
filelock==3.15.4
flatbuffers==24.3.25
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.65.2
h11==0.14.0
h5py==3.11.0
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
keras==3.4.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
maturin==1.7.0
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.4.0
mpmath==1.3.0
namex==0.0.8
networkx==3.3
numpy==1.26.4
opt-einsum==3.3.0
optree==0.12.1
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
protobuf==4.25.4
py-slippi==1.6.2
py-ubjson==0.16.1
pycparser==2.22
Pygments==2.18.0
Pymem==1.13.1
PySocks==1.7.1
random-user-agent==1.0.1
requests==2.32.3
rich==13.7.1
selenium==4.23.0
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sympy==1.13.1
tbb==2021.13.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
torch==2.3.1
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.2.2
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
wsproto==1.2.0
